requirements.txt
bs4
requests
schedule
python-dotenv